requirements.txt
nonebot
# requests
apscheduler
jieba
python-dateutil
msgpack
PyGithub